requirements.txt
pyYaml
b2sdk
